Funpay-GPT-Feedback/requirements.txt
g4f==0.5.7.6
python-dotenv
